stock_forecaster/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
statsmodels
yfinance
matplotlib
tensorflow
scikit-learn